- C#의 Encoding 클래스 C++에서 구현(+ icu 라이브러리 약간 개조..) - C#의 Random 클래스 C++에서 구현 - uint8_t -> std::byte로 교체 중 - 일부 클래스에 template deduction guide 추가 - boost 종속성 줄이려고 boost::noncopyable 제거, NonCopyable 추가

+++ Source/TGON/Core/RTTI.cpp
@@ -11,7 +11,7 @@ RTTI::RTTI(const std::type_info& typeInfo, const RTTI* superRTTI) noexcept :
 {
 }
 
-size_t RTTI::GetHashCode() const noexcept
+int32_t RTTI::GetHashCode() const noexcept
 {
     return m_typeInfo->hash_code();
 }
